+++ AquaMQTT/platformio.ini
@@ -13,8 +13,8 @@ test_framework = googletest
 # uncomment the next line if you have error when you upload the firmware (Error 74)
 #upload_protocol = esptool
 # uncomment the below lines to use over the air update
-#upload_protocol = espota
-#upload_port = 192.168.188.85
+upload_protocol = espota
+upload_port = 192.168.7.190
 lib_deps =
     locoduino/RingBuffer@1.0.5
     FrankBoesing/FastCRC@1.41.0
